requirements.txt
fastapi
uvicorn
httpx
aiogram
python-dotenv